Home_work_5S

+++ Program.cs
@@ -53,9 +53,9 @@ void newMass2(){
 void newMass3(){
     int length = new Random().Next(1, 10);
     Console.WriteLine($"массив newMass состоит из {length} чисел");
-    int[] newMass = new int[length];
-    int Pmax = newMass[0];
-    int Pmin = newMass[0];
+    double[] newMass = new double[length];
+    double Pmax = newMass[0];
+    double Pmin = newMass[0];
     for (int i = 0; i < length; i++)
     {
         newMass[i] = new Random().Next(100, 1000);
